feat: separate monthly column in it's own file and implemented the tooltip
