+++ lib/src/features/viewall/presentation/widgets/viewall_chart.dart
@@ -4,7 +4,7 @@ import 'package:skeletonizer/skeletonizer.dart';
 
 import '../../../../core/utils/utils.dart';
 import '../../domain/entities/chart.dart';
-import '../../domain/entities/monthly_chart_data.dart';
+import 'monthly_column.dart';
 
 class ViewallChart extends StatelessWidget {
   const ViewallChart({required this.chart, super.key});
@@ -176,124 +176,6 @@ class ViewallChartLoading extends StatelessWidget {
           ),
         ),
       ),
-    );
-  }
-}
-
-class MonthlyColumn extends StatefulWidget {
-  const MonthlyColumn({
-    required this.monthData,
-    required this.columnHeightFactor,
-    super.key,
-  });
-
-  final MonthlyChartDataEntity monthData;
-  final double columnHeightFactor;
-
-  @override
-  State<MonthlyColumn> createState() => _MonthlyColumnState();
-}
-
-class _MonthlyColumnState extends State<MonthlyColumn>
-    with TickerProviderStateMixin {
-  late final AnimationController _controller;
-  late final Animation<double> _animation;
-
-  @override
-  void initState() {
-    _controller = AnimationController(
-      duration: const Duration(milliseconds: 1),
-      vsync: this,
-    );
-    _animation = CurvedAnimation(
-      parent: _controller,
-      curve: Curves.ease,
-    );
-    _controller.forward();
-    super.initState();
-  }
-
-  @override
-  void dispose() {
-    _controller.dispose();
-    super.dispose();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      children: [
-        Expanded(
-          child: Container(
-            clipBehavior: Clip.antiAlias,
-            alignment: Alignment.bottomCenter,
-            width: 8,
-            decoration: BoxDecoration(
-              color: Colors.grey[200],
-              borderRadius: BorderRadius.circular(mediumSize),
-            ),
-            child: LayoutBuilder(
-              builder: (context, constraints) {
-                return AnimatedBuilder(
-                  animation: _animation,
-                  builder: (context, child) => AnimatedContainer(
-                    height: constraints.maxHeight *
-                        widget.columnHeightFactor *
-                        _animation.value,
-                    decoration: BoxDecoration(
-                      borderRadius: BorderRadius.circular(smallSize),
-                      gradient: buttonsGradient,
-                    ),
-                    curve: Curves.ease,
-                    duration: const Duration(milliseconds: 1300),
-                  ),
-                );
-              },
-            ),
-          ),
-        ),
-        const SizedBox(
-          height: 4,
-        ),
-        Text(
-          widget.monthData.date.month.toString().padLeft(2, '0'),
-          style: const TextStyle(
-            fontSize: smallText,
-          ),
-        ),
-      ],
-    );
-  }
-}
-
-class MonthlyColumnLoading extends StatelessWidget {
-  const MonthlyColumnLoading({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      children: [
-        Expanded(
-          child: Skeleton.leaf(
-            child: Container(
-              width: 8,
-              decoration: BoxDecoration(
-                borderRadius: BorderRadius.circular(mediumSize),
-                color: Colors.grey[200],
-              ),
-            ),
-          ),
-        ),
-        const SizedBox(
-          height: 4,
-        ),
-        const Text(
-          '01',
-          style: TextStyle(
-            fontSize: smallText,
-          ),
-        ),
-      ],
     );
   }
 }
